oskar_engine/requirements.txt
fastapi
uvicorn
torch
transformers
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
pydantic
sqlalchemy
psycopg2-binary
celery
redis
prometheus-client
torch-geometric
neo4j
spacy
openai-whisper
pytesseract
Pillow
langdetect
qdrant-client
python-multipart
